pyproject.toml
[project]
name = "boxhunt"
version = "0.1.0"
description = "A web scraper for collecting cardboard box images and a 3D box creation tool"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "requests>=2.31.0",
    "pillow>=10.0.0",
    "imagehash>=4.3.1",
    "pandas>=2.0.0",
    "aiohttp>=3.8.0",
    "asyncio-throttle>=1.0.2",
    "python-dotenv>=1.0.0",
    "tqdm>=4.65.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=4.9.0",
    # GUI and 3D dependencies
    "PySide6>=6.6.0",
    "PyOpenGL>=3.1.6",
    "PyOpenGL-accelerate>=3.1.6",
    "numpy>=1.24.0",
    "scipy>=1.11.0",
    "trimesh>=4.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py311"
line-length = 88

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff.lint.isort]
known-first-party = ["boxhunt"]

[dependency-groups]
dev = [
    "ruff>=0.12.8",
]

[project.scripts]
boxhunt = "boxhunt.main:main"
